requirements.txt
asgiref==3.8.1
beautifulsoup4>=4.12.0
brotli==1.2.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
fonttools==4.61.0
gunicorn==23.0.0
idna==3.10
joblib==1.5.2
msal==1.34.0
nltk==3.9.2
numpy==2.2.4
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.1.0
mssql-django>=1.5
pyodbc>=5.0.0
pycparser==2.23
pydyf==0.12.1
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
regex==2025.9.18
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
tzdata==2025.1
urllib3==2.3.0
weasyprint==67.0
webencodings==0.5.1
zopfli==0.4.0
openai>=1.0.0
celery[redis]>=5.4.0
redis>=5.0.0
azure-ai-documentintelligence>=1.0.0

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=5.0.0
pytest-timeout>=2.3.0
coverage>=7.0.0
